Moved project-creating code to its own module
